Modify Regex to support query params
Sometimes filenames contain query parameters in them, especially in css
`url` imports.

For example: If you try to import css from `font-awesome` none of the
fonts will load because the current regex fails to recognize that it
should match `../fonts/fontawesome-webfont.woff2?v=4.5.0` since the
regex only looks for files that end in `woff2`.
[https://github.com/FortAwesome/Font-Awesome/blob/master/css/font-awesome.css#L10](https://github.com/FortAwesome/Font-Awesome/blob/master/css/font-awesome.css#L10)

This will solve that problem by modifying the regex so that it will
match extensions with or without query parameters.